+++ projects/isy-angular-widgets/src/lib/i18n/widgets-config.service.ts
@@ -28,11 +28,12 @@ export class WidgetsConfigService {
     }
   };
 
-  private translationSource = new Subject<any>();
+  private translationSource = new Subject<unknown>();
 
-  getTranslation(path: string): any {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-member-access
-    return path.split('.').reduce((o, k) => (o ? (o as any)[k] : undefined), this.translation);
+  getTranslation(path: string): string | undefined {
+    // Needs to be refactored in the future
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-explicit-any
+    return path.split('.').reduce((o, k) => (o ? (o as any)[k] : undefined), this.translation) as string;
   }
 
   setTranslation(value: WidgetsTranslation): void {